Moving reconcile_users_org_team_mappings into common library
